Crate mdbook_renderer

Library to assist implementing an mdbook renderer.

This library is used to implement a renderer for mdBook. See the linked chapter for more information on how to implement a renderer.

Modules

book

A tree structure representing a book.

config

Mdbook’s configuration system.

errors

The error types used in mdbook.

Structs

RenderContext

The context provided to all renderers.

Constants

MDBOOK_VERSION

The current version of mdbook.

Traits

Renderer

An mdbook backend.